build(mcp): upgrade rmcp to 3.1
